Add MIT License and copyright to TwoRoomGameForm
The `TwoRoomGameForm` class was updated to include the MIT License
and copyright notice at the beginning of the file. The copyright
is attributed to "Joseph W. Lumbley" (2025). This change ensures
that the software's licensing terms are explicitly stated,
allowing for use, modification, and distribution under the
conditions of the MIT License.